Fixed projectile source filenames, indentations
