+++ src/Entities/FireChargeEntity.cpp
@@ -1,6 +1,6 @@
 #include "Globals.h"  // NOTE: MSVC stupidness requires this to be the same across all modules
 
-#include "ProjectileFireCharge.h"
+#include "FireChargeEntity.h"
 #include "../World.h"
 
 
@@ -8,7 +8,7 @@
 
 
 cFireChargeEntity::cFireChargeEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed) :
-super(pkFireCharge, a_Creator, a_X, a_Y, a_Z, 0.3125, 0.3125)
+	super(pkFireCharge, a_Creator, a_X, a_Y, a_Z, 0.3125, 0.3125)
 {
 	SetSpeed(a_Speed);
 	SetGravity(0);
