+++ src/Entities/FireChargeEntity.h
@@ -1,5 +1,5 @@
 //
-//  ProjectileGhastFireball.h
+//  FireChargeEntity.h
 //
 
 #pragma once
@@ -12,8 +12,8 @@
 
 // tolua_begin
 
-class cGhastFireballEntity :
-public cProjectileEntity
+class cFireChargeEntity :
+	public cProjectileEntity
 {
 	typedef cProjectileEntity super;
 	
@@ -21,9 +21,9 @@ public:
 	
 	// tolua_end
 	
-	CLASS_PROTODEF(cGhastFireballEntity);
+	CLASS_PROTODEF(cFireChargeEntity);
 	
-	cGhastFireballEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed);
+	cFireChargeEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed);
 	
 protected:
 	
@@ -32,7 +32,5 @@ protected:
 	// cProjectileEntity overrides:
 	virtual void OnHitSolidBlock(const Vector3d & a_HitPos, eBlockFace a_HitFace) override;
 	virtual void OnHitEntity(cEntity & a_EntityHit, const Vector3d & a_HitPos) override;
-	
-	// TODO: Deflecting the fireballs by arrow- or sword- hits
 	
 } ; // tolua_export
