+++ src/Entities/GhastFireballEntity.cpp
@@ -1,6 +1,6 @@
 #include "Globals.h"  // NOTE: MSVC stupidness requires this to be the same across all modules
 
-#include "ProjectileGhastFireball.h"
+#include "GhastFireballEntity.h"
 #include "../World.h"
 
 
@@ -8,7 +8,7 @@
 
 
 cGhastFireballEntity::cGhastFireballEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed) :
-super(pkGhastFireball, a_Creator, a_X, a_Y, a_Z, 1, 1)
+	super(pkGhastFireball, a_Creator, a_X, a_Y, a_Z, 1, 1)
 {
 	SetSpeed(a_Speed);
 	SetGravity(0);
