+++ src/Entities/ThrownEggEntity.cpp
@@ -1,6 +1,6 @@
 #include "Globals.h"  // NOTE: MSVC stupidness requires this to be the same across all modules
 
-#include "ProjectileEgg.h"
+#include "ThrownEggEntity.h"
 #include "../World.h"
 
 
@@ -8,7 +8,7 @@
 
 
 cThrownEggEntity::cThrownEggEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed) :
-super(pkEgg, a_Creator, a_X, a_Y, a_Z, 0.25, 0.25)
+	super(pkEgg, a_Creator, a_X, a_Y, a_Z, 0.25, 0.25)
 {
 	SetSpeed(a_Speed);
 }
