+++ src/Entities/ThrownEggEntity.h
@@ -1,5 +1,5 @@
 //
-//  ProjectileSnowball.h
+//  ThrownEggEntity.h
 //
 
 #pragma once
@@ -12,8 +12,8 @@
 
 // tolua_begin
 
-class cThrownSnowballEntity :
-public cProjectileEntity
+class cThrownEggEntity :
+	public cProjectileEntity
 {
 	typedef cProjectileEntity super;
 	
@@ -21,14 +21,17 @@ public:
 	
 	// tolua_end
 	
-	CLASS_PROTODEF(cThrownSnowballEntity);
+	CLASS_PROTODEF(cThrownEggEntity);
 	
-	cThrownSnowballEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed);
+	cThrownEggEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed);
 	
 protected:
 	
 	// cProjectileEntity overrides:
 	virtual void OnHitSolidBlock(const Vector3d & a_HitPos, eBlockFace a_HitFace) override;
 	virtual void OnHitEntity    (cEntity & a_EntityHit, const Vector3d & a_HitPos) override;
+	
+	// Randomly decides whether to spawn a chicken where the egg lands.
+	void TrySpawnChicken(const Vector3d & a_HitPos);
 	
 } ; // tolua_export
