+++ src/Entities/ThrownEnderPearlEntity.h
@@ -1,5 +1,5 @@
 //
-//  ProjectileEgg.h
+//  ThrownEnderPearlEntity.h
 //
 
 #pragma once
@@ -12,8 +12,8 @@
 
 // tolua_begin
 
-class cThrownEggEntity :
-public cProjectileEntity
+class cThrownEnderPearlEntity :
+	public cProjectileEntity
 {
 	typedef cProjectileEntity super;
 	
@@ -21,9 +21,9 @@ public:
 	
 	// tolua_end
 	
-	CLASS_PROTODEF(cThrownEggEntity);
+	CLASS_PROTODEF(cThrownEnderPearlEntity);
 	
-	cThrownEggEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed);
+	cThrownEnderPearlEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed);
 	
 protected:
 	
@@ -31,7 +31,7 @@ protected:
 	virtual void OnHitSolidBlock(const Vector3d & a_HitPos, eBlockFace a_HitFace) override;
 	virtual void OnHitEntity    (cEntity & a_EntityHit, const Vector3d & a_HitPos) override;
 	
-	// Randomly decides whether to spawn a chicken where the egg lands.
-	void TrySpawnChicken(const Vector3d & a_HitPos);
+	// Teleports the creator where the ender pearl lands.
+	void TeleportCreator(const Vector3d & a_HitPos);
 	
 } ; // tolua_export
