+++ src/Entities/ThrownSnowballEntity.cpp
@@ -1,6 +1,6 @@
 #include "Globals.h"  // NOTE: MSVC stupidness requires this to be the same across all modules
 
-#include "ProjectileSnowball.h"
+#include "ThrownSnowballEntity.h"
 #include "../World.h"
 
 
@@ -8,7 +8,7 @@
 
 
 cThrownSnowballEntity::cThrownSnowballEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed) :
-super(pkSnowball, a_Creator, a_X, a_Y, a_Z, 0.25, 0.25)
+	super(pkSnowball, a_Creator, a_X, a_Y, a_Z, 0.25, 0.25)
 {
 	SetSpeed(a_Speed);
 }
