+++ src/Entities/ThrownSnowballEntity.h
@@ -1,5 +1,5 @@
 //
-//  ProjectileEnderPearl.h
+//  ThrownSnowballEntity.h
 //
 
 #pragma once
@@ -12,8 +12,8 @@
 
 // tolua_begin
 
-class cThrownEnderPearlEntity :
-public cProjectileEntity
+class cThrownSnowballEntity :
+	public cProjectileEntity
 {
 	typedef cProjectileEntity super;
 	
@@ -21,17 +21,14 @@ public:
 	
 	// tolua_end
 	
-	CLASS_PROTODEF(cThrownEnderPearlEntity);
+	CLASS_PROTODEF(cThrownSnowballEntity);
 	
-	cThrownEnderPearlEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed);
+	cThrownSnowballEntity(cEntity * a_Creator, double a_X, double a_Y, double a_Z, const Vector3d & a_Speed);
 	
 protected:
 	
 	// cProjectileEntity overrides:
 	virtual void OnHitSolidBlock(const Vector3d & a_HitPos, eBlockFace a_HitFace) override;
 	virtual void OnHitEntity    (cEntity & a_EntityHit, const Vector3d & a_HitPos) override;
-	
-	// Teleports the creator where the ender pearl lands.
-	void TeleportCreator(const Vector3d & a_HitPos);
 	
 } ; // tolua_export
