+++ src/Items/ItemBow.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "../Entities/ProjectileArrow.h"
+#include "../Entities/ArrowEntity.h"
