+++ src/Mobs/Skeleton.cpp
@@ -3,7 +3,7 @@
 
 #include "Skeleton.h"
 #include "../World.h"
-#include "../Entities/ProjectileArrow.h"
+#include "../Entities/ArrowEntity.h"
