+++ src/Protocol/Protocol17x.cpp
@@ -31,8 +31,8 @@ Implements the 1.7.x protocol classes:
 #include "../BlockEntities/MobHeadEntity.h"
 #include "../BlockEntities/FlowerPotEntity.h"
 #include "../CompositeChat.h"
-#include "../Entities/ProjectileArrow.h"
-#include "../Entities/ProjectileFirework.h"
+#include "../Entities/ArrowEntity.h"
+#include "../Entities/FireworkEntity.h"
