+++ src/WorldStorage/NBTChunkSerializer.cpp
@@ -28,7 +28,7 @@
 #include "../Entities/Boat.h"
 #include "../Entities/Minecart.h"
 #include "../Entities/Pickup.h"
-#include "../Entities/ProjectileArrow.h"
+#include "../Entities/ArrowEntity.h"
 #include "../Entities/TNTEntity.h"
 #include "../Entities/ExpOrb.h"
 #include "../Entities/HangingEntity.h"
