+++ src/WorldStorage/WSSAnvil.cpp
@@ -35,12 +35,12 @@
 #include "../Entities/FallingBlock.h"
 #include "../Entities/Minecart.h"
 #include "../Entities/Pickup.h"
-#include "../Entities/ProjectileArrow.h"
-#include "../Entities/ProjectileEgg.h"
-#include "../Entities/ProjectileEnderPearl.h"
-#include "../Entities/ProjectileSnowball.h"
-#include "../Entities/ProjectileFireCharge.h"
-#include "../Entities/ProjectileGhastFireball.h"
+#include "../Entities/ArrowEntity.h"
+#include "../Entities/ThrownEggEntity.h"
+#include "../Entities/ThrownEnderPearlEntity.h"
+#include "../Entities/ThrownSnowballEntity.h"
+#include "../Entities/FireChargeEntity.h"
+#include "../Entities/GhastFireballEntity.h"
 #include "../Entities/TNTEntity.h"
 #include "../Entities/ExpOrb.h"
 #include "../Entities/HangingEntity.h"
